Refactor network packet chunk unpacking
Rename the `CNetRecvUnpacker` class to `CPacketChunkUnpacker` and add documentation to make its use clearer.

Rename the `FetchChunk` function to `UnpackNextChunk` to improve clarity. Return `bool` for success instead of `int`.

Rename the `Start` function to `FeedPacket` and also copy the packet explicity to improve clarity. Separate the buffer used for receiving packets in the `CNetClient::Recv` and `CNetServer::Recv` functions from the buffer used for unpacking chunks from a packet in the `CPacketChunkUnpacker` to improve clarity.

Remove the `Clear` function, as the chunk unpacker handles this internally and should not be cleared directly by users of the class.

Return early in the `UnpackNextChunk` function if the `CPacketChunkUnpacker` is not ready to unpack a chunk, as it is connection-oriented and cannot unpack chunks before the connection has been initialized and an appropriate non-control packet with at least one chunk was fed into it by calling the `FeedPacket` function.

Previously, all packets were first unpacked into the packet chunk unpacker's buffer and unpacking the chunks only failed gracefully because the non-packed connless and control packets had the number of chunks explicitly set to `0`, so the check `m_CurrentChunk >= m_Data.m_NumChunks` always caused these packets to be ignored by the chunk unpacker.

This fixes the ASAN `runtime error: index -1 out of bounds for type 'unsigned char [1394]'` on the line `unsigned char *pEnd = m_Data.m_aChunkData + m_Data.m_DataSize` when `m_Data.m_DataSize` is `-1` due to huffman decompression failing and the invalid packet being unpacked directly into the chunk unpacker's buffer. Closes #3346.

The checks whether `m_pConnection` is not `nullptr` are unnecessary, as the connection should always be initialized after the `FeedPacket` function has been called.