Fix `CTextCursor::m_MaxLines` not working with line breaks in text
The maximum number of lines was only handled correctly when text is wrapped automatically, but not when `\n` characters appear in the text. Previously, a `\n` character was effectively ignored if it exceeded the maximum number of lines, but subsequent non-space text was still rendered. Now, the text will be truncated when the maximum number of lines is exceeded due to a `\n` character.

(cherry picked from commit 8837a59425985976a407f1b1f095d89d3d35cda4)

# Conflicts:
#	src/engine/client/text.cpp