Fix fifo path in integration test
The fifo path in the integration test is relative to the project directory, but was passed to client/server whose working directory is a test-specific temporary folder. The fifo could therefore not be created, because the fifo path did not exist relative to the working directory.

Closes #11458.

(cherry picked from commit d0b51bdfa7516fce0fd02d5e42d8cd37fb4416b3)

# Conflicts:
#	scripts/integration_test.py