Fix macOS client always changing current working directory
This seems to be a hack to fix that launching the client with Finder on macOS caused the working directory to be incorrect. This was for example reported in https://stackoverflow.com/questions/4521151/mac-app-default-working-directory.

However, this was always overriding the current working directory with the folder of the executable even if the working directory was set correctly before launching the client.

The working directory is now changed only when the client is started by GUI.

The code already detects whether it is being launch by Finder by checking for the presence of the `-psn` (process serial number) argument, which is present for all apps launched by GUI.

(cherry picked from commit 00f16103db8d30e81834cfbaed537fc1d50d5638)